% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_candidate_motifs.R
\name{cluster_candidate_motifs}
\alias{cluster_candidate_motifs}
\title{cluster_candidate_motifs}
\usage{
cluster_candidate_motifs(
  filter_candidate_motifs_results,
  motif_overlap = 0.6,
  k_knn = 3,
  votes_knn_Rall = 0.5,
  votes_knn_Rm = 0.5,
  worker_number = NULL
)
}
\arguments{
\item{filter_candidate_motifs_results}{output of filter_candidate_motifs function.}

\item{motif_overlap}{minimum overlap required between candidate motifs in their distance computation,
in percentage of the shortest motif.}

\item{k_knn}{number of neighbors to be used in k-nearest neighbors, in order to determine Rall and Rm
when the two groups (curves with/without motif) are not separated.}

\item{votes_knn_Rall}{threshold on the percentage of votes for class 1 (curve has the motif) in
k-nearest neighbors, in order to determine Rall.}

\item{votes_knn_Rm}{threshold on the percentage of votes for class 1 (curve has the motif) in
k-nearest neighbors, in order to determine Rm.}

\item{worker_number}{number of CPU cores to be used for parallelization (default number of CPU cores).
If worker_number=1, the function is run sequentially.}
}
\value{
A list containing some of its input: k_knn, votes_knn_Rall and votes_knn_Rm, the outputs of filter_candidate_patterns function and:
  \item{VV_D}{ matrix of distances between candidates motifs after filtering}
  \item{VV_S}{ matrix of shift deformations of candidates motifs after filtering}
  \item{R_all}{ global radius used to cut the dendrogram}
  \item{hclust_res}{list of result hierarchical clustering based on the distances candidate motifs obtained after filtering}
}
\description{
Determine a global radius Rall (based on distances between all motifs and all curves), and cluster candidate
motifs based on their distance, requiring groups to be more than 2*Rall apart. Then for each group m=1,...,M
determine a group-specific radius Rm (based on distances between motifs of the same group and all curves).
}
\author{
Marzia Angela Cremona & Francesca Chiaromonte
}
