% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_candidate_motifs_plot.R
\name{cluster_candidate_motifs_plot}
\alias{cluster_candidate_motifs_plot}
\title{cluster_candidate_motifs_plot}
\usage{
cluster_candidate_motifs_plot(
  cluster_candidate_motifs_results,
  ylab = "",
  R_all = cluster_candidate_motifs_results$R_all,
  R_m = NULL,
  ask = TRUE
)
}
\arguments{
\item{cluster_candidate_motifs_results}{output of cluster_candidate_motifs function.}

\item{ylab}{Name of y axis}

\item{R_all}{global radius, used to cut the dendrogram (requiring groups to be more than 2*Rall apart).}

\item{R_m}{vector with group-specific radii. The length of the vector must match the number of clusters
obtained cutting the dendrogram at height 2*Rall. If NULL, Rm is determined in each group (based on
distances between motifs of the same group and all curves).}

\item{ask}{if TRUE the user is prompted before a new plot is drawn.}
}
\description{
Plot the results of cluster_candidate_motifs.
}
\author{
Marzia Angela Cremona & Francesca Chiaromonte
}
