% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diss_d0_d1_L2.R
\name{.diss_d0_d1_L2}
\alias{.diss_d0_d1_L2}
\title{.diss_d0_d1_L2}
\usage{
.diss_d0_d1_L2(y, v, w, alpha)
}
\arguments{
\item{y}{list of two elements y0=y(x), y1=y'(x) for x in dom(v), matrices with d columns.}

\item{v}{list of two elements v0=v(x), v1=v'(x) for x in dom(v), matrices with d columns.}

\item{w}{weights for the dissimilarity index in the different dimensions (w>0).}

\item{alpha}{weight coefficient between d0.L2 and d1.L2 (alpha=0 means d0.L2, alpha=1 means d1.L2).}
}
\description{
Dissimilarity index for multidimensional curves (dimension=d).
Sobolev type distance with normalization on common support: (1-alpha)*d0.L2+alpha*d1.L2.
}
\author{
Marzia Angela Cremona & Francesca Chiaromonte
}
