% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{.domain}
\alias{.domain}
\title{.domain}
\usage{
.domain(v, use0)
}
\arguments{
\item{v}{list of two elements: v0=v(x), v1=v'(x), matrices with d columns.}

\item{use0}{boolean: use0=TRUE, v0 is considered else v1 is considered.}
}
\value{
vector of boolean whose elements are TRUE iff at least one element in the row of v is not NA
}
\description{
Length of the domain of the curve without NA
}
\author{
Marzia Angela Cremona & Francesca Chiaromonte
}
