% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_min_diss.R
\name{.find_min_diss}
\alias{.find_min_diss}
\title{.find_min_diss}
\usage{
.find_min_diss(y, v, alpha, w, c_k, d, use0, use1)
}
\arguments{
\item{y}{list of two elements y0=y(x), y1=y'(x), matrices with d columns.}

\item{v}{list of two elements v0=v(x), v1=v'(x), matrices with d columns.}

\item{alpha}{weight coefficient between d0.L2 and d1.L2.}

\item{w}{weights for the dissimilarity index in the different dimensions (w>0).}

\item{c_k}{minimum length of supp(y_shifted) and supp(v) intersection.}
}
\value{
Shift warping and dissimilarity
}
\description{
Find shift warping minimizing dissimilarity between multidimensional curves (dimension=d).
}
\author{
Marzia Angela Cremona & Francesca Chiaromonte
}
