% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_occurrences.R
\name{.find_occurrences}
\alias{.find_occurrences}
\title{find_occurrences}
\usage{
.find_occurrences(v, Y, R, alpha, w, c_k, use0, use1, transformed = FALSE)
}
\arguments{
\item{v}{list of 2 elements, v0, v1, matrices with d columns.}

\item{Y}{list of N lists of two elements, Y0, Y1, matrices with d columns.}

\item{R}{maximum dissimilarity allowed.}

\item{alpha}{if diss_fun=diss_d0_d1_L2, weight coefficient between d0_L2 and d1_L2.}

\item{w}{weights for the dissimilarity index in the different dimensions (w>0).}

\item{c_k}{minimum length of supp(y_shifted) and supp(v) intersection.}
}
\value{
curve id, shift and dissimilarity
}
\description{
Find occurrences of a motif in a set of curves (dimesion=d), with dissimilarity lower than R.
}
\author{
Marzia Angela Cremona & Francesca Chiaromonte
}
