% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapply_custom.R
\name{.mapply_custom}
\alias{.mapply_custom}
\title{.mapply_custom}
\usage{
.mapply_custom(
  cl,
  FUN,
  ...,
  MoreArgs = NULL,
  SIMPLIFY = TRUE,
  USE.NAMES = TRUE
)
}
\description{
if the number of cluster is not null apply clusterMap, otherwise apply the classical version
}
\author{
Marzia Angela Cremona & Francesca Chiaromonte
}
