% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_domain.R
\name{.select_domain}
\alias{.select_domain}
\title{.select_domain}
\usage{
.select_domain(v, v_dom, use0, use1)
}
\arguments{
\item{v}{list of two elements: v0=v(x), v1=v'(x), matrices with d columns.}

\item{v_dom}{boolean vector: domain of v0 and v1.}

\item{use0}{boolean: if use0=TRUE, v0 is selected.}

\item{use1}{boolean: if use1=TRUE, v1 is selected.}
}
\value{
v list: containing the selected portion of v and v'.
}
\description{
Select the portion of a motif without NA
}
\author{
Marzia Angela Cremona  & Francesca Chiaromonte
}
