% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_candidate_motifs.R
\name{filter_candidate_motifs}
\alias{filter_candidate_motifs}
\title{filter_candidate_motifs}
\usage{
filter_candidate_motifs(
  find_candidate_motifs_results,
  sil_threshold = 0.5,
  size_threshold = 2,
  K = find_candidate_motifs_results$K,
  c = find_candidate_motifs_results$c
)
}
\arguments{
\item{find_candidate_motifs_results}{output of find_candidate_motif function.}

\item{sil_threshold}{threshold on the average silhouette index.}

\item{size_threshold}{threshold on the size of the motif (number of curves in the cluster).}

\item{K}{vector with numbers of motifs that must be considered (should be a subset of
find_candidate_motifs_results$K).}

\item{c}{vector with minimum motifs lengths that must be considered (should be a subset of
find_candidate_motifs_results$c).}
}
\value{
A list containing some ProbKMA options: Y0, Y1, diss, alpha, w, c, K and max_gap; and

\item{V0_clean}{ vector of candidates motifs after filtering}

\item{V1_clean}{ vector of deriveds candidates motifs after filtering}

\item{D_clean}{ matrix of distances candidates motifs from the curves after filtering}

\item{P_clean}{ matrix of probabilities membership of the candidate motifs after filtering}
}
\description{
Filter the candidate motifs on the basis of a threshold on the average silhouette index
and a threshold on the size of the curves in the motif.
}
\author{
Marzia Angela Cremona & Francesca Chiaromonte
}
