% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_methods.R
\name{generateCurves}
\alias{generateCurves}
\title{generateCurves}
\usage{
generateCurves(
  object,
  noise_type = NULL,
  noise_str = NULL,
  seed_background = 777,
  seed_motif = 43213,
  only_der = TRUE,
  coeff_min_shift = -10,
  coeff_max_shift = 10
)
}
\arguments{
\item{object}{An S4 object that has been previously constructed (mandatory).}

\item{noise_type}{A string specifying whether to add pointwise error or coefficients ('pointwise' or 'coeff') (mandatory).}

\item{noise_str}{A list corresponding to the number of motifs, specifying the structure of noise to be added for each motif. If 'pointwise' is chosen, the user can specify a list of vectors or matrices indicating the amount of noise for each motif. If 'coeff' is selected, a list of individual values or vectors can be provided (mandatory).}

\item{seed_background}{An integer specifying the seed for background curve generation (default = 777).}

\item{seed_motif}{An integer specifying the seed for motif generation (default = 43213).}

\item{only_der}{A logical value indicating whether to add a vertical shift to each motif instance. If set to 'FALSE', a vertical shift is added (default = TRUE).}

\item{coeff_min_shift}{A numeric value specifying the minimum vertical shift (default = -10).}

\item{coeff_max_shift}{A numeric value specifying the maximum vertical shift (default = 10).}
}
\value{
a list containing curves with embedded motifs.
}
\description{
Generate the curves with the motifs embedded.
}
\examples{
\dontrun{
curves <- mtfd::generateCurves(builder,noise_type = 'pointwise',noise_str)                                
}
}
