% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{initialChecks}
\alias{initialChecks}
\title{InitialChecks}
\usage{
initialChecks(Y0, Y1, P0, S0, params, diss, v_init)
}
\arguments{
\item{Y0}{curves}

\item{Y1}{derivatives}

\item{P0}{initial membership probabilities}

\item{S0}{initial shift matrix}

\item{params}{list of parameters}

\item{diss}{dissimilarity type}
}
\value{
\item{List}{returns a list containing FuncData and Parameters useful for initializing the probKMA object}
}
\description{
checks inputs provided by the user for running ProbKMA
}
