% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs_search_plot.R
\name{motifs_search_plot}
\alias{motifs_search_plot}
\title{motifs_search_plot}
\usage{
motifs_search_plot(
  motifs_search_results,
  ylab = "",
  freq_threshold = 5,
  top_n = "all",
  plot_curves = TRUE
)
}
\arguments{
\item{motifs_search_results}{output of motifs_search function.}

\item{ylab}{string with the title for the y axis .}

\item{freq_threshold}{plot only motifs with frequency at least equal to freq_threshold.}

\item{top_n}{if 'all', plot all motifs found. If top_n is an integer, then all top top_n motifs are plotted.}

\item{plot_curves}{if TRUE, plot all the curves with coloured motifs.}
}
\description{
Plot the results of motifs_search.
}
\author{
Marzia Angela Cremona & Francesca Chiaromonte
}
