% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_search_plot_norm.R
\name{motifs_search_plot_norm}
\alias{motifs_search_plot_norm}
\title{motifs_search_plot_norm}
\usage{
motifs_search_plot_norm(
  motifs_search_results,
  ylab = "",
  freq_threshold = 5,
  top_n = "all",
  plot_curves = TRUE,
  transformed = FALSE
)
}
\arguments{
\item{motifs_search_results}{output of motifs_search function.}

\item{ylab}{a vector of length d, with the titles for the y axis for each dimension.}

\item{freq_threshold}{plot only motifs with frequency at least equal to freq_threshold.}

\item{top_n}{if 'all', plot all motifs found. If top_n is an integer, then all top top_n motifs are plotted.}

\item{plot_curves}{if TRUE, plot all the curves with coloured motifs.}

\item{transformed}{if if TRUE, y is normalized to [0,1].}
}
\description{
Plot the results of motifs_search.
}
\author{
Marzia Angela Cremona & Francesca Chiaromonte
}
