% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{motifSimulationApp}
\alias{motifSimulationApp}
\title{motifSimulationApp}
\usage{
motifSimulationApp(noise_str, mot_details)
}
\arguments{
\item{noise_str}{A list corresponding to the number of motifs, specifying the structure of noise to be added for each motif. If 'pointwise' is chosen, the user can specify a list of vectors or matrices indicating the amount of noise for each motif. If 'coeff' is selected, a list of individual values or vectors can be provided.}

\item{mot_details}{A list outlining the definitions of the motifs to be included. Each motif is characterized by its length, a set of coefficients that may be optionally specified, and the number of occurrences. These occurrences can be indicated either by specific positions within the curves or by a total count. In the latter case, the algorithm will randomly position the motifs throughout the curves.}
}
\description{
Shiny app is available, serving as a graphical user interface (GUI) that enables users to execute all the motifSimulation functions in a straightforward and intuitive manner. The app consistently provides summary plots, enhancing the user experience.
}
\examples{
\dontrun{
mtfd::motifSimulationApp(noise_str,mot_details)
}
}
