% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtfd-pkg.R
\docType{package}
\name{mtfd-package}
\alias{mtfd}
\alias{mtfd-package}
\title{mtfd: Motif Discovery in Functional Data}
\description{
Method to locally cluster curves and discover functional motifs, i.e. typical “shapes” that may recur several times along and across the curves capturing important local characteristics. In order to identify these shared curve portions, such method leverages ideas from functional data analysis (joint clustering and alignment of curves), bioinformatics (local alignment through the extension of high similarity seeds) and fuzzy clustering (curves belonging to more than one cluster, if they contain more than one typical “shape”). It can employ various dissimilarity measures and incorporate derivatives in the discovery process, thus exploiting complex facets of shapes. Additionally, the package includes functionality to simulate functional data with embedded motifs. This feature allows users to create benchmark datasets to test the algorithms, ensuring robust validation and comparison of motif discovery methods.
}
