% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{padding}
\alias{padding}
\title{probKMA_plot}
\usage{
padding(dataMatrix, maxLen)
}
\arguments{
\item{dataMatrix}{matrix that must be extended}

\item{maxLen}{extension length}
}
\value{
matrix extended up to maxLen
}
\description{
Plot the results of probKMA.
}
\author{
Niccolò Feresini
}
