% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_methods.R
\name{plot_motifs}
\alias{plot_motifs}
\title{plot_motifs}
\usage{
plot_motifs(object, curves, name, path = getwd())
}
\arguments{
\item{object}{An S4 object that has been previously constructed (mandatory).}

\item{curves}{The result of the previous method (mandatory).}

\item{path}{A character string specifying the directory path where the results will be saved (mandatory).}
}
\description{
Plot the results of generateCurves.
}
\examples{
\dontrun{
mtfd::plot_motifs(builder,curves,getwd())                                  
}
}
