% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probKMA_plot.R
\name{probKMA_plot}
\alias{probKMA_plot}
\title{probKMA_plot}
\usage{
probKMA_plot(probKMA_results, ylab = "", cleaned = FALSE)
}
\arguments{
\item{probKMA_results}{output of probKMA function.}

\item{ylab}{a vector of length d, with the titles for the y axis for each dimension.}

\item{cleaned}{if TRUE, plot the cleaned motifs.}
}
\value{
plot of memberships motifs, objective function and Bhattacharyya distance
}
\description{
Plot the results of probKMA.
}
\author{
Marzia Angela Cremone & Francesca Chiaromonte
}
