% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probKMA_silhouette.R
\name{probKMA_silhouette}
\alias{probKMA_silhouette}
\title{probKMA_silhouette}
\usage{
probKMA_silhouette(probKMA_results, align = FALSE, plot = TRUE)
}
\arguments{
\item{probKMA_results}{output of probKMA function (with return_options=TRUE).}

\item{align}{if TRUE, try all possible alignements between pieces of curves (corresponding to the
same or to different motifs).}

\item{plot}{if TRUE, the silhouette plot is drawn.}
}
\value{
A list containing:

\item{silhouette}{ vector of silhouette indices}

\item{motifs}{ vector of motifs numbers}

\item{curves}{ vector of curves numbers with motifs}

\item{silhouette_average}{ vector of average silhouette index for each cluster}
}
\description{
Compute the adapted silhouette index on the results of probKMA.
}
\author{
Marzia Angela Cremona  & Francesca Chiaromonte
}
