% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probKMA_silhouette_filter.R
\name{probKMA_silhouette_filter}
\alias{probKMA_silhouette_filter}
\title{probKMA_silhouette_filter}
\usage{
probKMA_silhouette_filter(
  probKMA_results,
  silhouette,
  sil_threshold = 0.5,
  size_threshold = 2
)
}
\arguments{
\item{probKMA_results}{output of probKMA function (with return_options=TRUE)}

\item{silhouette}{output of probKMA_silhouette function}

\item{sil_threshold}{threshold on the average silhouette index}

\item{size_threshold}{threshold on the size of the motif (number of curves in the cluster)}
}
\value{
A list containing all the elements in probKMA_results correspondent to the filtered motifs;

\item{V0_clean}{ filtered motifs}

\item{V1_clean}{ derivative of the filtered motifs}

\item{D}{ filtered dissimilarity matrix}

\item{S_clean}{ shift warping matrix after cleaning motifs}

\item{D}{ dissimilarity matrix}

\item{D_clean}{ filtered dissimilarity matrix after cleaning motifs}

\item{P}{ filtered membership matrix}

\item{P_clean}{ filtered membership matrix after cleaning motifs}

\item{c}{ filtered minimum motif lengths}

\item{K}{ vector containing the number of motifs repeated a number of filtered motifs times}
}
\description{
Filter the motifs found by probKMA on the basis of a threshold on the average silhouette index
}
\author{
Marzia Angela Cremona & Francesca Chiaromonte
}
