% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probKMA_silhouette_plot.R
\name{probKMA_silhouette_plot}
\alias{probKMA_silhouette_plot}
\title{probKMA_silhouette}
\usage{
probKMA_silhouette_plot(silhouette_results, K, plot = TRUE)
}
\arguments{
\item{probKMA_results}{output of probKMA function (with return_options=TRUE).}
}
\value{
A list containing:

\item{silhouette}{ vector of silhouette indices}

\item{motifs}{ vector of motifs numbers}

\item{curves}{ vector of curves numbers with motifs}

\item{silhouette_average}{ vector of average silhouette index for each cluster}
}
\description{
Compute the adapted silhouette index on the results of probKMA.
}
\author{
Marzia Angela Cremona  & Francesca Chiaromonte
}
