% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_methods.R
\name{to_motifDiscovery}
\alias{to_motifDiscovery}
\title{to_motifDiscovery}
\usage{
to_motifDiscovery(curves)
}
\arguments{
\item{curves}{A list coming from generateCurves function.}
}
\value{
A  list containing all curves suitable for clusterMotif input.
}
\description{
Transforms the result of generateCurves into a format suitable for clusterMotif.
}
\examples{
\dontrun{
mtfd::plot(builder,curves,getwd())                                  
}
}
